+++ labfile.py
@@ -52,7 +52,7 @@ def coverage(test_files: List[Path]) -> None:
     test(test_files)
     cov.stop()
     cov.xml_report()
-    cov.report()
+    cov.report(show_missing=True)
 
 
 @alk.recipe(transient=True)
@@ -134,7 +134,6 @@ def release_test(build: Path) -> None:
 
     :param build: The build directory containing alkymi distributions to upload
     """
-    alk.utils.call(["python3", "-m", "pip", "install", "--user", "twine==3.2.0"])
     alk.utils.call(["python3", "-m", "twine", "upload", "--repository", "testpypi", "{}/*".format(build)])
 
 
@@ -145,7 +144,6 @@ def release(build: Path) -> None:
 
     :param build: The build directory containing alkymi distributions to upload
     """
-    alk.utils.call(["python3", "-m", "pip", "install", "--user", "twine==3.2.0"])
     alk.utils.call(["python3", "-m", "twine", "upload", "{}/*".format(build)])